refactor(github): split forgeProvider.ts into mappers, query infra, and mutations
- Extract pure GraphQL/REST-shape normalizers into mappers.ts
- Extract cache-key/dispatch/dedupe/singleflight infra into queryInfra.ts
- Extract list/get/find/CI-status read implementations into readOps.ts
- Extract create/edit/close/merge write implementations into mutations.ts,
  converting the object-literal methods into named exported functions
- Move REPO_METADATA_QUERY/PR_CI_STATUS_QUERY into GitHubQueries.ts
  alongside the rest of the GraphQL query constants
- forgeProvider.ts is now a thin composition root wiring the four
  modules plus the capability objects into githubForgeProvider
