+++ plugins/builtin/github/main/forgeProvider.ts
@@ -1,87 +1,34 @@
-import type { GraphQlQueryResponseData } from "@octokit/graphql";
 import { createHash } from "node:crypto";
-import { configure } from "safe-stable-stringify";
 import type {
   AuthValidation,
   AvatarCapability,
-  CIStatus,
-  CIStatusState,
-  CreateIssueInput,
-  CreatePRInput,
   Credentials,
-  EditPRInput,
-  EditIssueInput,
   FirstPageSnapshot,
   ForgeProviderImpl,
   ForgeRepoCounts,
   ForgeTokenHealthState,
   ForgeUser,
-  ForgeLabel,
   HealthEventsCapability,
   IdentityCapability,
-  Issue,
-  IssueCloseReason,
-  IssueComment,
   IssueTooltipData,
-  LinkedPRSummary,
-  ListOptions,
-  MergePRInput,
-  NormalizedIssueState,
-  NormalizedPRState,
-  NormalizedReviewDecision,
-  PR,
   PRTooltipData,
-  Page,
-  PRListProbeResult,
-  PRSnapshot,
   ProjectHealthCapability,
   ProjectHealthSnapshot,
-  PushErrorClassification,
   RateLimitDetails,
   RateLimitInfo,
-  RepoMetadata,
   RepoRef,
   RepoStatsCapability,
   RepoStatsSnapshot,
   ReviewCapability,
-  ReviewerRequest,
-  ReviewThread,
   StatsPage,
   TooltipCapability,
 } from "../../../../shared/types/forge.js";
-import { GitHubAuth, GITHUB_API_TIMEOUT_MS } from "./GitHubAuth.js";
+import { GitHubAuth } from "./GitHubAuth.js";
 import { cloneCapability } from "./GitHubClone.js";
 import { validateGitHubToken } from "./GitHubToken.js";
 import { parseGitHubRepoUrl } from "./GitHubRepoContext.js";
-import {
-  LIST_ISSUES_QUERY,
-  LIST_PRS_QUERY,
-  SEARCH_QUERY,
-  GET_ISSUE_QUERY,
-  GET_PR_QUERY,
-  GET_PR_REVIEW_THREADS_QUERY,
-  GET_PR_NODE_ID_QUERY,
-  CONVERT_PR_TO_DRAFT_MUTATION,
-  MARK_PR_READY_FOR_REVIEW_MUTATION,
-  BATCH_BRANCH_CHUNK_SIZE,
-  GRAPHQL_BATCH_CHUNK_SIZE,
-  buildBatchBranchPRQuery,
-  buildBatchPRsQuery,
-  buildBatchRequiredChecksQuery,
-} from "./GitHubQueries.js";
 import { gitHubRateLimitService } from "./GitHubRateLimitService.js";
-import {
-  clearGitHubCaches,
-  clearPRCaches,
-  forgeQueryCache,
-  forgeQueryInflight,
-  repoEventsETagCache,
-} from "./GitHubCaches.js";
-import { parseGitHubError } from "./GitHubErrors.js";
-import { probeOpenPRList } from "./GitHubPRDiscovery.js";
-import { deriveRequiredCIStatus } from "./prRequiredCIStatus.js";
-import { MAX_REVIEW_THREAD_PAGES } from "./GitHubCaches.js";
-import type { RollupContextNode } from "./prRequiredCIStatus.js";
+import { clearPRCaches, repoEventsETagCache } from "./GitHubCaches.js";
 import {
   fetchActivityProbe,
   getFirstPageCacheForContext,
@@ -89,1497 +36,55 @@ import {
 } from "./GitHubStats.js";
 import type { RepoStatsAndPageResult } from "./GitHubStats.js";
 import { getProjectHealthForContext } from "./GitHubHealth.js";
-import { getIssueTooltipForContext, getIssuesByNumbersForContext } from "./GitHubIssues.js";
+import { getIssueTooltipForContext } from "./GitHubIssues.js";
 import { getPRTooltipForContext } from "./GitHubPRs.js";
 import { resolveAuthorAvatar } from "./GitHubProfilePicture.js";
 import { gitHubTokenHealthService } from "./GitHubTokenHealthService.js";
 import { fetchRateLimitDetails } from "./GitHubRateLimitApi.js";
 import { toRateLimitInfo } from "./rateLimitUtils.js";
-import {
-  forgeIssueListCache,
-  forgePRListCache,
-  getRepoListEpoch,
-  issueTooltipCache,
-  prRequiredStatusCache,
-  truncateBody,
-  writePRTooltip,
-  type PRRequiredStatusEntry,
-} from "./GitHubCaches.js";
-import {
-  buildListCacheKey,
-  parseBatchRequiredChecksResponse,
-  updateRepoStatsCount,
-} from "./GitHubPRs.js";
-import type { GitHubIssue, GitHubPR, GitHubPRCIStatus, GitHubUser } from "../shared/types.js";
 import type { GitHubTokenHealthPayload } from "../shared/types.js";
-
-const REPO_METADATA_QUERY = `
-  query GetRepoMetadata($owner: String!, $repo: String!) {
-    repository(owner: $owner, name: $repo) {
-      defaultBranchRef { name }
-      isPrivate
-      isFork
-      isArchived
-      description
-      licenseInfo { name }
-      repositoryTopics(first: 20) { nodes { topic { name } } }
-    }
-    rateLimit { cost remaining resetAt limit }
-  }
-`;
-
-const PR_CI_STATUS_QUERY = `
-  query GetPRCIStatus($owner: String!, $repo: String!, $number: Int!) {
-    repository(owner: $owner, name: $repo) {
-      pullRequest(number: $number) {
-        commits(last: 1) {
-          nodes {
-            commit {
-              statusCheckRollup {
-                state
-                contexts(first: 100) {
-                  nodes {
-                    __typename
-                    ... on CheckRun {
-                      name
-                      conclusion
-                      status
-                      isRequired(pullRequestNumber: $number)
-                    }
-                    ... on StatusContext {
-                      context
-                      state
-                      isRequired(pullRequestNumber: $number)
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    rateLimit { cost remaining resetAt limit }
-  }
-`;
-
-function requireClient(): NonNullable<ReturnType<typeof GitHubAuth.createClient>> {
-  const client = GitHubAuth.createClient();
-  if (!client) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-  return client;
-}
-
-function isoToMs(value: unknown): number {
-  if (typeof value !== "string") return 0;
-  const t = Date.parse(value);
-  return Number.isFinite(t) ? t : 0;
-}
-
-function isoToMsOrNull(value: unknown): number | null {
-  if (typeof value !== "string") return null;
-  const t = Date.parse(value);
-  return Number.isFinite(t) ? t : null;
-}
-
-function toForgeUser(node: unknown): ForgeUser | undefined {
-  if (!node || typeof node !== "object") return undefined;
-  const n = node as { login?: unknown; avatarUrl?: unknown };
-  if (typeof n.login !== "string") return undefined;
-  return {
-    login: n.login,
-    ...(typeof n.avatarUrl === "string" ? { avatarUrl: n.avatarUrl } : {}),
-    rawData: node,
-  };
-}
-
-function toForgeUsers(node: unknown): ForgeUser[] {
-  const nodes = (node as { nodes?: unknown[] } | undefined)?.nodes;
-  if (!Array.isArray(nodes)) return [];
-  return nodes.map(toForgeUser).filter((u): u is ForgeUser => u !== undefined);
-}
-
-function toForgeLabels(node: unknown): ForgeLabel[] {
-  const nodes = (node as { nodes?: unknown[] } | undefined)?.nodes;
-  if (!Array.isArray(nodes)) return [];
-  const out: ForgeLabel[] = [];
-  for (const n of nodes) {
-    if (!n || typeof n !== "object") continue;
-    const label = n as { name?: unknown; color?: unknown };
-    if (typeof label.name !== "string") continue;
-    out.push({
-      name: label.name,
-      ...(typeof label.color === "string" ? { color: label.color } : {}),
-    });
-  }
-  return out;
-}
-
-function normalizeIssueState(rawState: string): NormalizedIssueState {
-  return rawState.toUpperCase() === "CLOSED" ? "closed" : "open";
-}
-
-function normalizePRState(rawState: string, merged: boolean): NormalizedPRState {
-  if (merged) return "merged";
-  const upper = rawState.toUpperCase();
-  if (upper === "CLOSED") return "closed";
-  if (upper === "MERGED") return "merged";
-  return "open";
-}
-
-function toForgeIssue(node: Record<string, unknown>): Issue {
-  const rawState = typeof node.state === "string" ? node.state : "OPEN";
-  const comments = node.comments as { totalCount?: unknown } | undefined;
-  const commentCount = typeof comments?.totalCount === "number" ? comments.totalCount : undefined;
-  return {
-    number: node.number as number,
-    title: (node.title as string) ?? "",
-    body: (node.bodyText as string) ?? "",
-    state: normalizeIssueState(rawState),
-    rawState,
-    url: (node.url as string) ?? "",
-    author: toForgeUser(node.author),
-    assignees: toForgeUsers(node.assignees),
-    labels: toForgeLabels(node.labels),
-    ...(commentCount !== undefined ? { commentCount } : {}),
-    createdAt: isoToMs(node.createdAt ?? node.updatedAt),
-    updatedAt: isoToMs(node.updatedAt),
-    closedAt: isoToMsOrNull(node.closedAt),
-    rawData: node,
-  };
-}
-
-/**
- * Map a GitHub REST user object ({@link https://docs.github.com/en/rest/issues})
- * to a {@link ForgeUser}. REST uses `avatar_url` where the GraphQL projection
- * (consumed by {@link toForgeUser}) uses `avatarUrl`, so REST responses need
- * their own mapper.
- */
-function restUserToForgeUser(node: unknown): ForgeUser | undefined {
-  if (!node || typeof node !== "object") return undefined;
-  const n = node as { login?: unknown; avatar_url?: unknown };
-  if (typeof n.login !== "string") return undefined;
-  return {
-    login: n.login,
-    ...(typeof n.avatar_url === "string" ? { avatarUrl: n.avatar_url } : {}),
-    rawData: node,
-  };
-}
-
-/**
- * Normalize a GitHub REST issue payload (e.g. the `POST .../issues` create
- * response) to the contract {@link Issue}. The REST field names diverge from
- * the GraphQL ones {@link toForgeIssue} handles — `html_url` vs `url`, `body`
- * vs `bodyText`, snake_case timestamps, and flat `labels`/`assignees` arrays
- * rather than `{ nodes: [...] }` connections — so a dedicated mapper is needed.
- */
-function restToForgeIssue(raw: Record<string, unknown>): Issue {
-  const rawState = typeof raw.state === "string" ? raw.state : "open";
-  const assignees = Array.isArray(raw.assignees)
-    ? raw.assignees.map(restUserToForgeUser).filter((u): u is ForgeUser => u !== undefined)
-    : [];
-  const labels: ForgeLabel[] = [];
-  if (Array.isArray(raw.labels)) {
-    for (const entry of raw.labels) {
-      if (typeof entry === "string") {
-        labels.push({ name: entry });
-        continue;
-      }
-      if (!entry || typeof entry !== "object") continue;
-      const label = entry as { name?: unknown; color?: unknown };
-      if (typeof label.name !== "string") continue;
-      labels.push({
-        name: label.name,
-        ...(typeof label.color === "string" ? { color: label.color } : {}),
-      });
-    }
-  }
-  return {
-    number: raw.number as number,
-    title: typeof raw.title === "string" ? raw.title : "",
-    body: typeof raw.body === "string" ? raw.body : "",
-    state: normalizeIssueState(rawState),
-    rawState,
-    url: typeof raw.html_url === "string" ? raw.html_url : "",
-    author: restUserToForgeUser(raw.user),
-    assignees,
-    labels,
-    createdAt: isoToMs(raw.created_at ?? raw.updated_at),
-    updatedAt: isoToMs(raw.updated_at),
-    closedAt: isoToMsOrNull(raw.closed_at),
-    rawData: raw,
-  };
-}
-
-/**
- * Normalize a GitHub REST pull-request payload (the `POST .../pulls` create and
- * `PATCH .../pulls/{n}` edit responses) to the contract {@link PR}. REST field
- * names diverge from the GraphQL projection {@link toForgePR} handles — `body`
- * vs `bodyText`, `draft` vs `isDraft`, snake_case timestamps, nested
- * `base.ref`/`head.ref` rather than `baseRefName`/`headRefName` — so a dedicated
- * mapper is needed.
- */
-function restToForgePR(raw: Record<string, unknown>): PR {
-  const merged = raw.merged === true || typeof raw.merged_at === "string";
-  const rawState = typeof raw.state === "string" ? raw.state : "open";
-  const base = raw.base as { ref?: unknown } | undefined;
-  const head = raw.head as { ref?: unknown } | undefined;
-  const mergeable = typeof raw.mergeable === "boolean" ? raw.mergeable : null;
-  return {
-    number: raw.number as number,
-    title: typeof raw.title === "string" ? raw.title : "",
-    body: typeof raw.body === "string" ? raw.body : "",
-    state: normalizePRState(rawState, merged),
-    rawState,
-    isDraft: raw.draft === true,
-    merged,
-    url: typeof raw.html_url === "string" ? raw.html_url : "",
-    author: restUserToForgeUser(raw.user),
-    baseRef: typeof base?.ref === "string" ? base.ref : "",
-    headRef: typeof head?.ref === "string" ? head.ref : "",
-    mergeable,
-    createdAt: isoToMs(raw.created_at ?? raw.updated_at),
-    updatedAt: isoToMs(raw.updated_at),
-    closedAt: isoToMsOrNull(raw.closed_at),
-    mergedAt: isoToMsOrNull(raw.merged_at),
-    rawData: raw,
-  };
-}
-
-/**
- * Map a GitHub REST label array (the body returned by the add/remove-label
- * endpoints) to the contract {@link ForgeLabel}[]. Entries without a string
- * `name` are skipped.
- */
-function restToForgeLabels(raw: unknown): ForgeLabel[] {
-  if (!Array.isArray(raw)) return [];
-  const labels: ForgeLabel[] = [];
-  for (const entry of raw) {
-    if (!entry || typeof entry !== "object") continue;
-    const label = entry as { name?: unknown; color?: unknown };
-    if (typeof label.name !== "string") continue;
-    labels.push({
-      name: label.name,
-      ...(typeof label.color === "string" ? { color: label.color } : {}),
-    });
-  }
-  return labels;
-}
-
-function requireGitHubToken(): string {
-  const token = GitHubAuth.getToken();
-  if (!token) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-  return token;
-}
-
-function githubMutationHeaders(token: string): Record<string, string> {
-  return {
-    Authorization: `Bearer ${token}`,
-    Accept: "application/vnd.github+json",
-    "X-GitHub-Api-Version": "2022-11-28",
-    "Content-Type": "application/json",
-  };
-}
-
-/**
- * Shared `PATCH /repos/{owner}/{repo}/issues/{number}` call backing
- * close/reopen/edit. Sends the given partial issue fields, validates the
- * mutation result, clears the issue caches, and returns the normalized
- * {@link Issue}. Callers own building the body (state/state_reason/title/body).
- */
-async function patchIssue(
-  repo: RepoRef,
-  issueNumber: number,
-  fields: Record<string, unknown>,
-  failurePrefix: string
-): Promise<Issue> {
-  const token = requireGitHubToken();
-  const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}`;
-  const response = await fetch(url, {
-    method: "PATCH",
-    headers: githubMutationHeaders(token),
-    body: JSON.stringify(fields),
-    signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    throw new Error(
-      `${failurePrefix}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-    );
-  }
-  const data = (await response.json()) as Record<string, unknown>;
-  if (typeof data.number !== "number" || typeof data.html_url !== "string") {
-    throw new Error("Unexpected response from GitHub: missing issue number or URL.");
-  }
-  // Provider-owned invalidation: state/title/body changes affect list and stat
-  // views, so drop the caches (mirrors createIssue).
-  clearGitHubCaches();
-  return restToForgeIssue(data);
-}
-
-// Shared PATCH for close/reopen — both flip `state` on the same endpoint.
-async function patchPRState(
-  repo: RepoRef,
-  prNumber: number,
-  state: "open" | "closed"
-): Promise<void> {
-  const token = requireGitHubToken();
-  const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}`;
-  const response = await fetch(url, {
-    method: "PATCH",
-    headers: githubMutationHeaders(token),
-    body: JSON.stringify({ state }),
-    signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    const verb = state === "closed" ? "close" : "reopen";
-    throw new Error(
-      `Failed to ${verb} pull request #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-    );
-  }
-}
-
-// Resolve a PR's GraphQL node id for the draft-toggle mutations.
-async function fetchPRNodeId(repo: RepoRef, prNumber: number): Promise<string> {
-  const response = await dispatchQuery(
-    GET_PR_NODE_ID_QUERY,
-    { owner: repo.owner, repo: repo.repo, number: prNumber },
-    "GET_PR_NODE_ID_QUERY"
-  );
-  const id = (response as { repository?: { pullRequest?: { id?: unknown } | null } | null })
-    ?.repository?.pullRequest?.id;
-  if (typeof id !== "string" || !id) {
-    throw new Error(`Pull request #${prNumber} not found.`);
-  }
-  return id;
-}
-
-function toForgePR(node: Record<string, unknown>): PR {
-  const merged = node.merged === true;
-  const rawState = typeof node.state === "string" ? node.state : "OPEN";
-  const commits = node.commits as
-    | { nodes?: Array<{ commit?: { statusCheckRollup?: { state?: unknown } | null } | null }> }
-    | undefined;
-  const rollupState = commits?.nodes?.[0]?.commit?.statusCheckRollup?.state;
-  const ciStatus = mapListCIStatus(
-    typeof rollupState === "string" ? (rollupState as GitHubPRCIStatus) : undefined
-  );
-  const comments = node.comments as { totalCount?: unknown } | undefined;
-  const commentCount = typeof comments?.totalCount === "number" ? comments.totalCount : undefined;
-  return {
-    number: node.number as number,
-    title: (node.title as string) ?? "",
-    body: (node.bodyText as string) ?? "",
-    state: normalizePRState(rawState, merged),
-    rawState,
-    isDraft: node.isDraft === true,
-    merged,
-    url: (node.url as string) ?? "",
-    author: toForgeUser(node.author),
-    baseRef: (node.baseRefName as string) ?? "",
-    headRef: (node.headRefName as string) ?? "",
-    mergeable: undefined,
-    reviewDecision: node.reviewDecision as NormalizedReviewDecision | undefined,
-    ...(commentCount !== undefined ? { commentCount } : {}),
-    ...(ciStatus ? { ciStatus } : {}),
-    createdAt: isoToMs(node.createdAt ?? node.updatedAt),
-    updatedAt: isoToMs(node.updatedAt),
-    closedAt: isoToMsOrNull(node.closedAt),
-    mergedAt: isoToMsOrNull(node.mergedAt),
-    rawData: node,
-  };
-}
-
-/**
- * GitHub list-shape CI roll-up → forge {@link CIStatusState}. `ERROR` folds
- * into `failure` and `EXPECTED` into `pending` — the forge set is boolean-ish
- * by design. `undefined` when the list shape carried no roll-up.
- */
-function mapListCIStatus(raw: GitHubPRCIStatus | undefined): CIStatusState | undefined {
-  if (raw === "SUCCESS") return "success";
-  if (raw === "FAILURE" || raw === "ERROR") return "failure";
-  if (raw === "PENDING" || raw === "EXPECTED") return "pending";
-  return undefined;
-}
-
-function gitHubUserToForgeUser(user: GitHubUser): ForgeUser {
-  return { login: user.login, avatarUrl: user.avatarUrl, rawData: null };
-}
-
-/**
- * Map a legacy {@link GitHubIssue} list item (the stats first-page shape) to
- * the contract {@link Issue}. The list shape carries no `body`/`createdAt`, so
- * `body` falls back to `""` and `createdAt` to the item's `updatedAt` —
- * matching {@link toForgeIssue}'s `createdAt ?? updatedAt` fallback.
- */
-export function gitHubIssueToForgeIssue(item: GitHubIssue): Issue {
-  const updatedAt = isoToMs(item.updatedAt);
-  const linkedCI = mapListCIStatus(item.linkedPR?.ciStatus);
-  const linkedPR: LinkedPRSummary | undefined = item.linkedPR
-    ? {
-        number: item.linkedPR.number,
-        state: normalizePRState(item.linkedPR.state, item.linkedPR.state === "MERGED"),
-        url: item.linkedPR.url,
-        ...(linkedCI ? { ciStatus: linkedCI } : {}),
-      }
-    : undefined;
-  return {
-    number: item.number,
-    title: item.title,
-    body: "",
-    state: normalizeIssueState(item.state),
-    rawState: item.state,
-    url: item.url,
-    author: gitHubUserToForgeUser(item.author),
-    assignees: item.assignees.map(gitHubUserToForgeUser),
-    labels: (item.labels ?? []).map((l) => ({ name: l.name, color: l.color })),
-    commentCount: item.commentCount,
-    ...(linkedPR ? { linkedPR } : {}),
-    createdAt: updatedAt,
-    updatedAt,
-    rawData: item,
-  };
-}
-
-/**
- * Map a legacy {@link GitHubPR} list item (the stats first-page shape) to the
- * contract {@link PR}. Same `body`/`createdAt` fallbacks as
- * {@link gitHubIssueToForgeIssue}; `baseRef` is absent from the list shape and
- * falls back to `""`.
- */
-export function gitHubPRToForgePR(item: GitHubPR): PR {
-  const updatedAt = isoToMs(item.updatedAt);
-  const merged = item.state === "MERGED";
-  const ciStatus = mapListCIStatus(item.ciStatus);
-  return {
-    number: item.number,
-    title: item.title,
-    body: item.bodyText ?? "",
-    state: normalizePRState(item.state, merged),
-    rawState: item.state,
-    isDraft: item.isDraft,
-    merged,
-    url: item.url,
-    author: gitHubUserToForgeUser(item.author),
-    baseRef: "",
-    headRef: item.headRefName ?? "",
-    mergeable: undefined,
-    reviewDecision: item.reviewDecision,
-    ...(item.commentCount !== undefined ? { commentCount: item.commentCount } : {}),
-    ...(ciStatus ? { ciStatus } : {}),
-    createdAt: updatedAt,
-    updatedAt,
-    rawData: item,
-  };
-}
-
-function mapIssueGraphQLStates(state: ListOptions["state"]): string[] {
-  if (state === "closed") return ["CLOSED"];
-  if (state === "all") return ["OPEN", "CLOSED"];
-  return ["OPEN"];
-}
-
-function mapPRGraphQLStates(state: ListOptions["state"]): string[] {
-  if (state === "merged") return ["MERGED"];
-  if (state === "closed") return ["CLOSED", "MERGED"];
-  if (state === "all") return ["OPEN", "CLOSED", "MERGED"];
-  return ["OPEN"];
-}
-
-function buildOrderBy(opts: ListOptions): { field: string; direction: string } {
-  const direction = opts.direction === "asc" ? "ASC" : "DESC";
-  const field = opts.sort === "updated" ? "UPDATED_AT" : "CREATED_AT";
-  return { field, direction };
-}
-
-// Deterministic stringify so equivalent variables produce one cache key
-// regardless of property insertion order across call sites.
-const stringifyVariables = configure({ bigint: false });
-
-// Hash the cache inputs to a fixed-size digest rather than storing the full
-// multi-KB query document as the key. `queryLabel` and `variables` distinguish
-// the named queries; the document is still folded in because batch queries
-// (BATCH_BRANCH_PR / BATCH_PRS / BATCH_REQUIRED_CHECKS) carry their distinct
-// branch/PR set in the document itself with `variables === {}` — without it
-// they'd all collide on one key. `\0` can't appear in a GraphQL document or
-// queryLabel, so neither field can be forged by the serialized variables.
-function buildCacheKey(
-  query: string,
-  variables: Record<string, unknown>,
-  queryLabel: string
-): string {
-  return createHash("sha256")
-    .update(`${queryLabel}\0${stringifyVariables(variables) ?? ""}\0${query}`)
-    .digest("hex");
-}
-
-async function dispatchQuery(
-  query: string,
-  variables: Record<string, unknown>,
-  queryLabel: string
-): Promise<GraphQlQueryResponseData> {
-  const client = requireClient();
-  try {
-    const response = (await client(query, {
-      ...variables,
-      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
-    })) as GraphQlQueryResponseData;
-    gitHubRateLimitService.updateFromGraphQL(response, queryLabel);
-    return response;
-  } catch (error) {
-    throw new Error(parseGitHubError(error), { cause: error });
-  }
-}
-
-/**
- * Sole GraphQL entry point. Serves a 60s response cache and coalesces
- * concurrent identical queries through an in-flight singleflight map (both in
- * `GitHubCaches.ts` so a token change clears them atomically). Errors are never
- * cached — a transient failure must not block retries for the full TTL.
- */
-async function runQuery(
-  query: string,
-  variables: Record<string, unknown>,
-  queryLabel: string,
-  bypass = false
-): Promise<GraphQlQueryResponseData> {
-  const key = buildCacheKey(query, variables, queryLabel);
-
-  if (!bypass) {
-    const cached = forgeQueryCache.get(key);
-    if (cached !== undefined) return cached;
-
-    const inflight = forgeQueryInflight.get(key);
-    if (inflight !== undefined) return inflight;
-  }
-
-  const request = dispatchQuery(query, variables, queryLabel)
-    .then((response) => {
-      // `bypass` skips the raw-response cache write too: the list paths
-      // (LIST_ISSUES/LIST_PRS) hold the strictly-larger raw response only to
-      // build a normalized `Page<T>` they cache separately, so caching the raw
-      // form here is redundant. Reads/inflight-joins are already bypassed above.
-      if (!bypass) forgeQueryCache.set(key, response);
-      return response;
-    })
-    .finally(() => {
-      if (forgeQueryInflight.get(key) === request) forgeQueryInflight.delete(key);
-    });
-
-  forgeQueryInflight.set(key, request);
-  return request;
-}
-
-/**
- * Single-flight coalescing maps for higher-level provider methods. These layer
- * on top of `runQuery`'s GraphQL-level dedup to coalesce at the call-site
- * boundary too — `listPRs`/`getCIStatus` etc. carry their own cache shape
- * (Page<T>, CIStatus) and need to dedup on the same key the cache uses, not the
- * underlying GraphQL key. Concurrent calls with the same key join the one
- * in-flight request instead of each paying full query cost — the dominant load
- * source is the worktree dashboard's fleet-wide PR/CI poll firing the same
- * lookups from every view at once.
- */
-const listIssuesInflight = new Map<string, Promise<Page<Issue>>>();
-const listPRsInflight = new Map<string, Promise<Page<PR>>>();
-const getIssueInflight = new Map<string, Promise<Issue | null>>();
-const getPRInflight = new Map<string, Promise<PR | null>>();
-const getCIStatusInflight = new Map<string, Promise<CIStatus | null>>();
-const findPRsByBranchesInflight = new Map<string, Promise<Map<string, PR | null>>>();
-const findPRsByNumbersInflight = new Map<string, Promise<Map<number, PR | null>>>();
-const getCIStatusesInflight = new Map<string, Promise<Map<number, CIStatus | null>>>();
-
-/**
- * Join an in-flight request for `key` when one exists, else start `fn` and
- * register it. `bypass` forces a fresh request (skips the join) and installs
- * itself as the new shared promise so callers arriving mid-flight get the fresh
- * result, not the stale one. Cleanup removes the entry only if it's still the
- * active promise, so a bypass replacement isn't deleted by the request it
- * superseded. Failures evict immediately so a transient error doesn't pin a
- * rejected promise for later callers.
- *
- * `fn` receives an `isCurrent()` guard: it returns `true` only while this call
- * is still the active in-flight entry. A request that was superseded by a newer
- * bypass call sees `false` and must skip any shared-cache write, so a slow
- * stale fetch can't overwrite the fresher result the bypass already committed.
- */
-function dedupe<T>(
-  inflight: Map<string, Promise<T>>,
-  key: string,
-  bypass: boolean,
-  fn: (isCurrent: () => boolean) => Promise<T>
-): Promise<T> {
-  if (!bypass) {
-    const pending = inflight.get(key);
-    if (pending) return pending;
-  }
-  const holder: { promise: Promise<T> | null } = { promise: null };
-  // Defer `fn` one microtask so `holder.promise` is assigned before it runs;
-  // `isCurrent` can then compare against this call's own promise identity.
-  const promise = Promise.resolve().then(() => fn(() => inflight.get(key) === holder.promise));
-  holder.promise = promise;
-  inflight.set(key, promise);
-  const cleanup = () => {
-    if (inflight.get(key) === promise) inflight.delete(key);
-  };
-  promise.then(cleanup, cleanup);
-  return promise;
-}
-
-function listCacheState(opts: ListOptions): "open" | "closed" | "merged" | "all" {
-  return opts.state ?? "open";
-}
-
-function issueToTooltipData(issue: Issue): IssueTooltipData {
-  return {
-    number: issue.number,
-    title: issue.title,
-    bodyExcerpt: truncateBody(issue.body),
-    state: issue.state,
-    rawState: issue.rawState,
-    createdAt: issue.createdAt,
-    author: {
-      login: issue.author?.login ?? "unknown",
-      avatarUrl: issue.author?.avatarUrl ?? "",
-      rawData: null,
-    },
-    assignees: issue.assignees.map((a) => ({
-      login: a.login,
-      avatarUrl: a.avatarUrl ?? "",
-      rawData: null,
-    })),
-    labels: issue.labels.map((l) => ({ name: l.name, color: l.color ?? "" })),
-  };
-}
-
-/**
- * Build a {@link PRTooltipData} from a normalized forge {@link PR}. Assignees
- * and labels aren't on the forge `PR` shape, so they're read from `rawData`
- * (present when the source query selected them — `GET_PR_QUERY` and
- * `buildBatchBranchPRQuery` both do). Returns `null` when `createdAt` is
- * missing (0), since the tooltip renders a real creation date.
- */
-function prToTooltipData(pr: PR): PRTooltipData | null {
-  if (!pr.createdAt) return null;
-  const raw = (pr.rawData ?? {}) as Record<string, unknown>;
-  const assigneeNodes =
-    (raw.assignees as { nodes?: Array<{ login?: string; avatarUrl?: string }> } | undefined)
-      ?.nodes ?? [];
-  const labelNodes =
-    (raw.labels as { nodes?: Array<{ name?: string; color?: string }> } | undefined)?.nodes ?? [];
-  return {
-    number: pr.number,
-    title: pr.title,
-    bodyExcerpt: truncateBody(pr.body),
-    state: pr.merged ? "merged" : pr.state,
-    rawState: pr.rawState,
-    isDraft: pr.isDraft,
-    createdAt: pr.createdAt,
-    author: {
-      login: pr.author?.login ?? "unknown",
-      avatarUrl: pr.author?.avatarUrl ?? "",
-      rawData: null,
-    },
-    assignees: assigneeNodes
-      .filter(Boolean)
-      .map((a) => ({ login: a.login ?? "unknown", avatarUrl: a.avatarUrl ?? "", rawData: null })),
-    labels: labelNodes.filter(Boolean).map((l) => ({ name: l.name ?? "", color: l.color ?? "" })),
-  };
-}
-
-function buildCIStatus(entry: PRRequiredStatusEntry, rawData: unknown): CIStatus {
-  let state: CIStatus["state"] = "unknown";
-  const effective = entry.ciStatus;
-  if (effective === "SUCCESS") state = "success";
-  else if (effective === "FAILURE" || effective === "ERROR") state = "failure";
-  else if (effective === "PENDING" || effective === "EXPECTED") state = "pending";
-  else if (effective === undefined && (entry.ciSummary?.requiredTotal ?? 0) === 0)
-    state = "neutral";
-
-  const total = entry.ciSummary?.requiredTotal ?? 0;
-  const failed = entry.ciSummary?.requiredFailing ?? 0;
-  const pending = entry.ciSummary?.requiredPending ?? 0;
-  const passed = Math.max(0, total - failed - pending);
-  const requiredChecksPassing =
-    entry.ciSummary !== undefined
-      ? entry.ciSummary.requiredTotal > 0 &&
-        entry.ciSummary.requiredFailing === 0 &&
-        entry.ciSummary.requiredPending === 0
-      : undefined;
-
-  return {
-    state,
-    total,
-    passed,
-    failed,
-    pending,
-    ...(requiredChecksPassing !== undefined ? { requiredChecksPassing } : {}),
-    rawData,
-  };
-}
-
-function uniquePositiveIntegers(values: number[]): number[] {
-  const unique: number[] = [];
-  const seen = new Set<number>();
-  for (const value of values) {
-    if (!Number.isInteger(value) || value <= 0 || seen.has(value)) continue;
-    seen.add(value);
-    unique.push(value);
-  }
-  return unique;
-}
-
-function buildNumberBatchInflightKey(repo: RepoRef, numbers: number[]): string {
-  return `${repo.owner}/${repo.repo}:${[...numbers].sort((a, b) => a - b).join(",")}`;
-}
-
-/**
- * Text-search path for `listIssues` — routes through GitHub's search API
- * instead of the repository issues connection. Results are typed-input
- * ephemera: they use a `search:`-prefixed in-flight dedupe key and are never
- * written to `forgeIssueListCache`, so a search response can't be served
- * later as the unfiltered background-poll list. `runQuery`'s short-TTL
- * response cache still coalesces identical search terms.
- */
-async function searchIssuesImpl(
-  repo: RepoRef,
-  search: string,
-  opts: ListOptions
-): Promise<Page<Issue>> {
-  const state = listCacheState(opts);
-  const bypass = opts.bypassCache === true;
-  const limit = opts.perPage ?? 20;
-  // Free text is appended unquoted — GitHub search tokenizes it as keywords,
-  // matching what its own search box does (see findPRByBranchImpl, which only
-  // quotes because branch refs must not parse as separate operators). GitHub
-  // caps search queries at 256 chars, so the term is truncated to the budget
-  // left after the qualifiers rather than letting the whole query be rejected.
-  const stateQualifier = state === "all" ? "" : ` state:${state}`;
-  const sortQualifier = opts.sort === "updated" ? "sort:updated-desc" : "sort:created-desc";
-  const prefix = `repo:${repo.owner}/${repo.repo} is:issue${stateQualifier} ${sortQualifier} `;
-  const available = 256 - prefix.length;
-  const searchQuery = `${prefix}${available > 0 ? search.slice(0, available) : ""}`.trim();
-  const dedupeKey = `search:${searchQuery}:${opts.cursor ?? ""}:${limit}`;
-
-  return dedupe(listIssuesInflight, dedupeKey, bypass, async () => {
-    const response = await runQuery(
-      SEARCH_QUERY,
-      {
-        searchQuery,
-        type: "ISSUE",
-        cursor: opts.cursor ?? null,
-        limit,
-      },
-      "SEARCH_QUERY",
-      bypass
-    );
-
-    const result = response?.search as
-      | {
-          nodes?: unknown[];
-          pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-          issueCount?: number;
-        }
-      | undefined;
-    const nodes = (result?.nodes ?? []) as Array<Record<string, unknown>>;
-    return {
-      items: nodes.filter(Boolean).map(toForgeIssue),
-      nextCursor: result?.pageInfo?.endCursor ?? null,
-      hasMore: result?.pageInfo?.hasNextPage ?? false,
-      ...(typeof result?.issueCount === "number" ? { totalCount: result.issueCount } : {}),
-    };
-  });
-}
-
-async function listIssuesImpl(repo: RepoRef, opts: ListOptions): Promise<Page<Issue>> {
-  const searchTerm = opts.search?.trim();
-  if (searchTerm) return searchIssuesImpl(repo, searchTerm, opts);
-
-  const state = listCacheState(opts);
-  const sortOrder = opts.sort === "updated" ? "updated" : "created";
-  const bypass = opts.bypassCache === true;
-  // The unfiltered list path keeps `search` out of the cache key — search
-  // routes through `searchIssuesImpl` above and never touches this cache.
-  const cacheKey = buildListCacheKey(
-    "issue",
-    repo.owner,
-    repo.repo,
-    state,
-    "",
-    sortOrder,
-    opts.cursor ?? ""
-  );
-
-  if (!bypass) {
-    const cached = forgeIssueListCache.get(cacheKey);
-    if (cached) return cached;
-  }
-
-  return dedupe(listIssuesInflight, cacheKey, bypass, async (isCurrent) => {
-    const limit = opts.perPage ?? 20;
-    const orderBy = buildOrderBy(opts);
-    // Captured before the network call: if the count-as-cache-buster bumps
-    // the epoch mid-flight, this page predates the observed change and must
-    // not repopulate the just-busted cache.
-    const epochAtStart = getRepoListEpoch("issue", repo.owner, repo.repo);
-
-    const response = await runQuery(
-      LIST_ISSUES_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        states: mapIssueGraphQLStates(opts.state),
-        cursor: opts.cursor ?? null,
-        limit,
-        orderBy,
-      },
-      "LIST_ISSUES_QUERY",
-      // Always bypass the raw-response cache: the normalized `Page<Issue>` is
-      // cached in `forgeIssueListCache` and the raw form would be redundant
-      // (and strictly larger). This path has its own `dedupe` above, so the
-      // runQuery-level singleflight isn't needed either.
-      true
-    );
-
-    const issues = (response?.repository as Record<string, unknown> | undefined)?.issues as
-      | {
-          nodes?: unknown[];
-          pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-          totalCount?: number;
-        }
-      | undefined;
-    const nodes = (issues?.nodes ?? []) as Array<Record<string, unknown>>;
-    const page: Page<Issue> = {
-      items: nodes.filter(Boolean).map(toForgeIssue),
-      nextCursor: issues?.pageInfo?.endCursor ?? null,
-      hasMore: issues?.pageInfo?.hasNextPage ?? false,
-      ...(typeof issues?.totalCount === "number" ? { totalCount: issues.totalCount } : {}),
-    };
-
-    // Skip the shared-cache write when a newer bypass call has superseded us
-    // or the count buster invalidated this repo's issue pages mid-flight, so
-    // a slow stale fetch can't clobber the fresher committed result. The
-    // stats-count update sits behind the same epoch guard: a response too old
-    // to repopulate the list cache is also too old to roll `repoStatsCache`
-    // back to its pre-change total under a fresh `lastUpdated`.
-    if (isCurrent() && getRepoListEpoch("issue", repo.owner, repo.repo) === epochAtStart) {
-      forgeIssueListCache.set(cacheKey, page);
-      if (state === "open" && !opts.cursor && typeof issues?.totalCount === "number") {
-        updateRepoStatsCount(`${repo.owner}/${repo.repo}`, "issue", issues.totalCount);
-      }
-    }
-    return page;
-  });
-}
-
-async function listPRsImpl(repo: RepoRef, opts: ListOptions): Promise<Page<PR>> {
-  const state = listCacheState(opts);
-  const sortOrder = opts.sort === "updated" ? "updated" : "created";
-  const bypass = opts.bypassCache === true;
-  // The PR list query ignores `opts.search` (advisory — see ListOptions), so
-  // it's kept out of the cache key. Wiring it would mean routing to
-  // SEARCH_QUERY like `searchIssuesImpl` does for issues.
-  const cacheKey = buildListCacheKey(
-    "pr",
-    repo.owner,
-    repo.repo,
-    state,
-    "",
-    sortOrder,
-    opts.cursor ?? ""
-  );
-
-  if (!bypass) {
-    const cached = forgePRListCache.get(cacheKey);
-    if (cached) return cached;
-  }
-
-  return dedupe(listPRsInflight, cacheKey, bypass, async (isCurrent) => {
-    const limit = opts.perPage ?? 20;
-    const orderBy = buildOrderBy(opts);
-    // Same mid-flight count-buster guard as the issues list above.
-    const epochAtStart = getRepoListEpoch("pr", repo.owner, repo.repo);
-
-    const response = await runQuery(
-      LIST_PRS_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        states: mapPRGraphQLStates(opts.state),
-        cursor: opts.cursor ?? null,
-        limit,
-        orderBy,
-      },
-      "LIST_PRS_QUERY",
-      // Always bypass the raw-response cache: the normalized `Page<PR>` is
-      // cached in `forgePRListCache` and the raw form would be redundant (and
-      // strictly larger). This path has its own `dedupe` above, so the
-      // runQuery-level singleflight isn't needed either.
-      true
-    );
-
-    const prs = (response?.repository as Record<string, unknown> | undefined)?.pullRequests as
-      | {
-          nodes?: unknown[];
-          pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-          totalCount?: number;
-        }
-      | undefined;
-    const nodes = (prs?.nodes ?? []) as Array<Record<string, unknown>>;
-    const page: Page<PR> = {
-      items: nodes.filter(Boolean).map(toForgePR),
-      nextCursor: prs?.pageInfo?.endCursor ?? null,
-      hasMore: prs?.pageInfo?.hasNextPage ?? false,
-      ...(typeof prs?.totalCount === "number" ? { totalCount: prs.totalCount } : {}),
-    };
-
-    if (isCurrent() && getRepoListEpoch("pr", repo.owner, repo.repo) === epochAtStart) {
-      forgePRListCache.set(cacheKey, page);
-      if (state === "open" && !opts.cursor && typeof prs?.totalCount === "number") {
-        updateRepoStatsCount(`${repo.owner}/${repo.repo}`, "pr", prs.totalCount);
-      }
-    }
-    return page;
-  });
-}
-
-async function getIssueImpl(repo: RepoRef, number: number): Promise<Issue | null> {
-  const key = `${repo.owner}/${repo.repo}:${number}`;
-  return dedupe(getIssueInflight, key, false, async () => {
-    const response = await runQuery(
-      GET_ISSUE_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        number,
-      },
-      "GET_ISSUE_QUERY"
-    );
-    const issue = (response?.repository as Record<string, unknown> | undefined)?.issue as
-      | Record<string, unknown>
-      | null
-      | undefined;
-    if (!issue) return null;
-    const forgeIssue = toForgeIssue(issue);
-    // Warm the hover-tooltip cache as a side effect so a subsequent hover skips
-    // a redundant fetch. The detail fetch can't *read* this cache — it holds
-    // the narrower tooltip shape, not a full Issue.
-    issueTooltipCache.set(key, issueToTooltipData(forgeIssue));
-    return forgeIssue;
-  });
-}
-
-async function getPRImpl(repo: RepoRef, number: number): Promise<PR | null> {
-  const key = `${repo.owner}/${repo.repo}:${number}`;
-  return dedupe(getPRInflight, key, false, async () => {
-    const requestedAt = Date.now();
-    const response = await runQuery(
-      GET_PR_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        number,
-      },
-      "GET_PR_QUERY"
-    );
-    const pr = (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
-      | Record<string, unknown>
-      | null
-      | undefined;
-    if (!pr) return null;
-    const forgePR = toForgePR(pr);
-    // Side-effect tooltip pre-warm (write-through only); guarded by the
-    // ownership-token timestamp so a slow response can't clobber a fresher one.
-    const tooltip = prToTooltipData(forgePR);
-    if (tooltip) writePRTooltip(repo.owner, repo.repo, number, tooltip, requestedAt);
-    return forgePR;
-  });
-}
-
-async function findPRsByBranchesImpl(
-  repo: RepoRef,
-  branches: string[]
-): Promise<Map<string, PR | null>> {
-  if (branches.length === 0) return new Map<string, PR | null>();
-
-  // Deduplicate while preserving order. Duplicates in the input still land
-  // in the result via the same key, but each unique value is queried once.
-  const unique: string[] = [];
-  const seen = new Set<string>();
-  for (const branch of branches) {
-    if (!seen.has(branch)) {
-      seen.add(branch);
-      unique.push(branch);
-    }
-  }
-
-  // Coalesce concurrent fleet-wide sweeps requesting the same branch set
-  // (order-independent) into one round-trip. The result Map preserves each
-  // caller's branch keys regardless of the sorted key used for coalescing.
-  const inflightKey = `${repo.owner}/${repo.repo}:${[...unique].sort().join(",")}`;
-  return dedupe(findPRsByBranchesInflight, inflightKey, false, async () => {
-    const result = new Map<string, PR | null>();
-
-    // Skip the network entirely while rate-limited — the caller's missing-key
-    // fallback handles every omitted branch. Mirrors `batchCheckLinkedPRs`.
-    const block = gitHubRateLimitService.shouldBlockRequest("graphql");
-    if (block.blocked) return result;
-
-    const requestedAt = Date.now();
-    // Chunks run sequentially (parallel would spike GraphQL points during a
-    // fleet-wide refresh). Per-chunk failures are caught so a single transient
-    // chunk error doesn't blank every branch — the caller falls back per-branch
-    // for any branch the result Map omits.
-    for (let start = 0; start < unique.length; start += BATCH_BRANCH_CHUNK_SIZE) {
-      const chunk = unique.slice(start, start + BATCH_BRANCH_CHUNK_SIZE);
-      const query = buildBatchBranchPRQuery(repo.owner, repo.repo, chunk);
-      let response: Record<string, unknown>;
-      try {
-        response = (await runQuery(query, {}, "BATCH_BRANCH_PR_QUERY")) as Record<string, unknown>;
-      } catch {
-        // Omit this chunk's branches from the result — caller's missing-key
-        // fallback path handles them.
-        continue;
-      }
-
-      for (let i = 0; i < chunk.length; i++) {
-        const branch = chunk[i];
-        // A missing alias key (vs. a present key with empty nodes) indicates a
-        // partial GraphQL response. Omit so the caller routes to fallback rather
-        // than silently recording "no PR found".
-        if (!(`b${i}` in response)) continue;
-        const aliasNode = response[`b${i}`] as
-          | { pullRequests?: { nodes?: unknown[] } }
-          | null
-          | undefined;
-        if (aliasNode == null) continue;
-        const nodes = (aliasNode.pullRequests?.nodes ?? []) as Array<Record<string, unknown>>;
-        const first = nodes.find(Boolean);
-        const forgePR = first ? toForgePR(first) : null;
-        result.set(branch, forgePR);
-
-        // Pre-warm the PR tooltip cache so a hover right after a fleet refresh
-        // is instant. The batch-branch query carries assignees/labels, so the
-        // warmed entry is complete (not a partial that a later hover refetches).
-        if (forgePR) {
-          const tooltip = prToTooltipData(forgePR);
-          if (tooltip) writePRTooltip(repo.owner, repo.repo, forgePR.number, tooltip, requestedAt);
-        }
-      }
-    }
-
-    return result;
-  });
-}
-
-async function findPRByBranchImpl(repo: RepoRef, branchName: string): Promise<PR | null> {
-  // Quote the branch name so refs containing spaces or characters that would
-  // otherwise be parsed as a separate search operator (`sort:`, `head:`,
-  // `is:`) don't override the intended search semantics.
-  const escapedBranch = branchName.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
-  const searchQuery = `repo:${repo.owner}/${repo.repo} is:pr head:"${escapedBranch}" sort:created-desc`;
-  const response = await runQuery(
-    SEARCH_QUERY,
-    {
-      searchQuery,
-      type: "ISSUE",
-      cursor: null,
-      limit: 1,
-    },
-    "SEARCH_QUERY"
-  );
-  const nodes = ((response?.search as { nodes?: unknown[] } | undefined)?.nodes ?? []) as Array<
-    Record<string, unknown>
-  >;
-  const first = nodes.find(Boolean);
-  return first ? toForgePR(first) : null;
-}
-
-async function findPRsByNumbersImpl(
-  repo: RepoRef,
-  prNumbers: number[]
-): Promise<Map<number, PR | null>> {
-  const unique = uniquePositiveIntegers(prNumbers);
-  if (unique.length === 0) return new Map<number, PR | null>();
-
-  const inflightKey = buildNumberBatchInflightKey(repo, unique);
-  return dedupe(findPRsByNumbersInflight, inflightKey, false, async () => {
-    const result = new Map<number, PR | null>();
-
-    const block = gitHubRateLimitService.shouldBlockRequest("graphql");
-    if (block.blocked) return result;
-
-    const requestedAt = Date.now();
-    for (let start = 0; start < unique.length; start += GRAPHQL_BATCH_CHUNK_SIZE) {
-      const chunk = unique.slice(start, start + GRAPHQL_BATCH_CHUNK_SIZE);
-      const query = buildBatchPRsQuery(repo.owner, repo.repo, chunk);
-      if (!query) continue;
-
-      let response: GraphQlQueryResponseData;
-      try {
-        response = await runQuery(query, {}, "BATCH_PRS_QUERY");
-      } catch {
-        // Omit this chunk so callers treat it as transient and keep stale data.
-        continue;
-      }
-
-      const repository = response?.repository as Record<string, unknown> | undefined;
-      if (!repository) continue;
-
-      for (const prNumber of chunk) {
-        const alias = `p${prNumber}`;
-        if (!(alias in repository)) continue;
-        const node = repository[alias] as Record<string, unknown> | null | undefined;
-        const forgePR = node ? toForgePR(node) : null;
-        result.set(prNumber, forgePR);
-
-        if (forgePR) {
-          const tooltip = prToTooltipData(forgePR);
-          if (tooltip) writePRTooltip(repo.owner, repo.repo, forgePR.number, tooltip, requestedAt);
-        }
-      }
-    }
-
-    return result;
-  });
-}
-
-async function getCIStatusesImpl(
-  repo: RepoRef,
-  prNumbers: number[]
-): Promise<Map<number, CIStatus | null>> {
-  const unique = uniquePositiveIntegers(prNumbers);
-  if (unique.length === 0) return new Map<number, CIStatus | null>();
-
-  const inflightKey = buildNumberBatchInflightKey(repo, unique);
-  return dedupe(getCIStatusesInflight, inflightKey, false, async () => {
-    const result = new Map<number, CIStatus | null>();
-    const missing: number[] = [];
-
-    for (const prNumber of unique) {
-      const cacheKey = `${repo.owner}/${repo.repo}:${prNumber}`;
-      const cached = prRequiredStatusCache.get(cacheKey);
-      if (cached) {
-        result.set(prNumber, buildCIStatus(cached, null));
-      } else {
-        missing.push(prNumber);
-      }
-    }
-
-    if (missing.length === 0) return result;
-
-    const block = gitHubRateLimitService.shouldBlockRequest("graphql");
-    if (block.blocked) return result;
-
-    for (let start = 0; start < missing.length; start += GRAPHQL_BATCH_CHUNK_SIZE) {
-      const chunk = missing.slice(start, start + GRAPHQL_BATCH_CHUNK_SIZE);
-      const query = buildBatchRequiredChecksQuery(repo.owner, repo.repo, chunk);
-      if (!query) continue;
-
-      let response: GraphQlQueryResponseData;
-      try {
-        response = await runQuery(query, {}, "BATCH_REQUIRED_CHECKS_QUERY");
-      } catch {
-        // Omit this chunk so callers treat it as transient and retry later.
-        continue;
-      }
-
-      const responseRecord = response as Record<string, unknown>;
-      const fetched = parseBatchRequiredChecksResponse(responseRecord, chunk);
-      for (const prNumber of chunk) {
-        const alias = `pr_${prNumber}`;
-        if (!(alias in responseRecord)) continue;
-        const repoNode = responseRecord[alias] as { pullRequest?: unknown } | null | undefined;
-        if (repoNode == null || repoNode.pullRequest == null) {
-          result.set(prNumber, null);
-          continue;
-        }
-        const entry = fetched.get(prNumber) ?? {
-          ciStatus: undefined,
-          ciSummary: undefined,
-        };
-        prRequiredStatusCache.set(`${repo.owner}/${repo.repo}:${prNumber}`, entry);
-        result.set(prNumber, buildCIStatus(entry, null));
-      }
-    }
-
-    return result;
-  });
-}
-
-async function getCIStatusImpl(repo: RepoRef, prNumber: number): Promise<CIStatus | null> {
-  // Shares the 60s required-status cache with the legacy enrich path (same key
-  // and {@link PRRequiredStatusEntry} value). This is the hottest forge path —
-  // the worktree dashboard polls CI per open PR across every view — so the
-  // cache + single-flight collapse the fan-out to one request per PR per 60s.
-  const cacheKey = `${repo.owner}/${repo.repo}:${prNumber}`;
-  const cached = prRequiredStatusCache.get(cacheKey);
-  if (cached) return buildCIStatus(cached, null);
-
-  return dedupe(getCIStatusInflight, cacheKey, false, async () => {
-    const response = await runQuery(
-      PR_CI_STATUS_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        number: prNumber,
-      },
-      "PR_CI_STATUS_QUERY"
-    );
-
-    const pr = (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
-      | Record<string, unknown>
-      | null
-      | undefined;
-    if (!pr) return null;
-
-    const commits = pr.commits as
-      | { nodes?: Array<{ commit?: { statusCheckRollup?: unknown } }> }
-      | undefined;
-    const rollup = commits?.nodes?.[0]?.commit?.statusCheckRollup as
-      | {
-          state?: string;
-          contexts?: { nodes?: RollupContextNode[]; pageInfo?: { hasNextPage?: boolean } };
-        }
-      | undefined;
-
-    const contextNodes = rollup?.contexts?.nodes ?? null;
-    const hasNextPage = rollup?.contexts?.pageInfo?.hasNextPage === true;
-    const derived = deriveRequiredCIStatus(contextNodes, hasNextPage, rollup?.state ?? null);
-
-    const entry: PRRequiredStatusEntry = {
-      ciStatus: derived.ciStatus,
-      ciSummary: derived.ciSummary,
-    };
-    prRequiredStatusCache.set(cacheKey, entry);
-    return buildCIStatus(entry, rollup ?? null);
-  });
-}
-
-async function probeOpenPRListImpl(
-  repo: RepoRef,
-  tracked: PRSnapshot[]
-): Promise<PRListProbeResult> {
-  const token = GitHubAuth.getToken();
-  // No token → the conditional GET can't be made; let the caller revalidate
-  // through its normal (also token-gated) path rather than claim "unchanged".
-  if (!token) return { kind: "fallback" };
-  return probeOpenPRList(repo.owner, repo.repo, token, tracked);
-}
-
-async function getRepoMetadataImpl(repo: RepoRef): Promise<RepoMetadata> {
-  const response = await runQuery(
-    REPO_METADATA_QUERY,
-    {
-      owner: repo.owner,
-      repo: repo.repo,
-    },
-    "REPO_METADATA_QUERY"
-  );
-
-  const repository = (response?.repository as Record<string, unknown> | undefined) ?? {};
-  const defaultBranch =
-    ((repository.defaultBranchRef as { name?: unknown } | null | undefined)?.name as
-      | string
-      | undefined) ?? "main";
-  const license =
-    ((repository.licenseInfo as { name?: unknown } | null | undefined)?.name as string | null) ??
-    null;
-  const topicNodes = ((
-    repository.repositoryTopics as { nodes?: Array<{ topic?: { name?: unknown } }> } | undefined
-  )?.nodes ?? []) as Array<{ topic?: { name?: unknown } }>;
-  const topics = topicNodes
-    .map((n) => (typeof n.topic?.name === "string" ? n.topic.name : null))
-    .filter((s): s is string => s !== null);
-
-  return {
-    defaultBranch,
-    isPrivate: repository.isPrivate === true,
-    isFork: repository.isFork === true,
-    isArchived: repository.isArchived === true,
-    description: (repository.description as string | null | undefined) ?? null,
-    license,
-    topics,
-    rawData: repository,
-  };
-}
-
-async function getReviewThreadsImpl(repo: RepoRef, prNumber: number): Promise<ReviewThread[]> {
-  const threads: ReviewThread[] = [];
-  let cursor: string | null = null;
-  let pageCount = 0;
-  while (true) {
-    const response = await runQuery(
-      GET_PR_REVIEW_THREADS_QUERY,
-      {
-        owner: repo.owner,
-        repo: repo.repo,
-        number: prNumber,
-        cursor,
-      },
-      "GET_PR_REVIEW_THREADS_QUERY"
-    );
-    const reviewThreads = (
-      (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
-        | {
-            reviewThreads?: {
-              nodes?: unknown[];
-              pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
-            };
-          }
-        | undefined
-    )?.reviewThreads;
-    const nodes = (reviewThreads?.nodes ?? []) as Array<Record<string, unknown>>;
-    for (const n of nodes) {
-      if (!n) continue;
-      const id = `${repo.owner}/${repo.repo}#${prNumber}:${threads.length}`;
-      threads.push({ id, rawData: n });
-    }
-    pageCount++;
-    if (pageCount >= MAX_REVIEW_THREAD_PAGES) {
-      break;
-    }
-    if (reviewThreads?.pageInfo?.hasNextPage && reviewThreads.pageInfo.endCursor) {
-      cursor = reviewThreads.pageInfo.endCursor;
-      continue;
-    }
-    break;
-  }
-  return threads;
-}
-
-function getRateLimitImpl(): Promise<RateLimitInfo> {
-  const state = gitHubRateLimitService.getState();
-  if (!state.blocked) {
-    return Promise.resolve({
-      limit: null,
-      remaining: null,
-      resetAt: null,
-      throttleMultiplier: state.throttleMultiplier ?? 1,
-    });
-  }
-  return Promise.resolve({
-    limit: null,
-    remaining: 0,
-    resetAt: state.resetAt ?? null,
-    ...(state.kind === "secondary" ? { secondaryThrottled: true } : {}),
-    throttleMultiplier: state.throttleMultiplier ?? 1,
-  });
-}
-
-/**
- * Extracts a `GH###` code (e.g. `GH006`, `GH013`) from raw push stderr — these
- * are GitHub's stable, googleable identifiers for protected-branch and ruleset
- * rejections. Returns `null` when no recognized code is present so the banner
- * falls back to its generic state.
- */
-function classifyPushErrorImpl(stderr: string): PushErrorClassification | null {
-  const match = /\bGH\d{3,}\b/.exec(stderr);
-  return match ? { code: match[0] } : null;
-}
-
-const REVIEW_WRITE_HEADERS = (token: string): Record<string, string> => ({
-  Authorization: `Bearer ${token}`,
-  Accept: "application/vnd.github+json",
-  "X-GitHub-Api-Version": "2022-11-28",
-  "Content-Type": "application/json",
-});
-
-/**
- * Submit a review verdict (`APPROVE` / `REQUEST_CHANGES`) on a PR. GitHub
- * returns 200 on success and 422 for refusals like approving your own PR — the
- * response body is surfaced so the caller sees the forge's own reason.
- */
-async function submitReviewImpl(
-  repo: RepoRef,
-  prNumber: number,
-  event: "APPROVE" | "REQUEST_CHANGES",
-  body?: string
-): Promise<void> {
-  const token = GitHubAuth.getToken();
-  if (!token) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-  const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}/reviews`;
-  const payload: Record<string, unknown> = { event };
-  if (body !== undefined) payload.body = body;
-  const response = await fetch(url, {
-    method: "POST",
-    headers: REVIEW_WRITE_HEADERS(token),
-    body: JSON.stringify(payload),
-    signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    const verb = event === "APPROVE" ? "approve" : "request changes on";
-    throw new Error(
-      `Failed to ${verb} PR #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-    );
-  }
-  // A submitted verdict changes the PR's `reviewDecision`, which is part of the
-  // cached PR object — invalidate so the next `getPR` reflects it (#9061).
-  clearPRCaches();
-}
-
-async function dismissReviewImpl(
-  repo: RepoRef,
-  prNumber: number,
-  reviewId: number,
-  message: string
-): Promise<void> {
-  const token = GitHubAuth.getToken();
-  if (!token) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-  const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}/reviews/${reviewId}/dismissals`;
-  const response = await fetch(url, {
-    method: "PUT",
-    headers: REVIEW_WRITE_HEADERS(token),
-    body: JSON.stringify({ message, event: "DISMISS" }),
-    signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    throw new Error(
-      `Failed to dismiss review ${reviewId} on PR #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-    );
-  }
-  clearPRCaches();
-}
-
-async function requestReviewersImpl(
-  repo: RepoRef,
-  prNumber: number,
-  reviewers: ReviewerRequest
-): Promise<void> {
-  const token = GitHubAuth.getToken();
-  if (!token) {
-    throw new Error("GitHub token not configured. Set it in Settings.");
-  }
-  const users = reviewers.users ?? [];
-  const teams = reviewers.teams ?? [];
-  if (users.length === 0 && teams.length === 0) {
-    throw new Error("Provide at least one user or team to request a review from");
-  }
-  const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}/requested_reviewers`;
-  const response = await fetch(url, {
-    method: "POST",
-    headers: REVIEW_WRITE_HEADERS(token),
-    // GitHub keys teams separately as `team_reviewers` (team slugs).
-    body: JSON.stringify({ reviewers: users, team_reviewers: teams }),
-    signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    throw new Error(
-      `Failed to request reviewers on PR #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-    );
-  }
-}
+import { gitHubIssueToForgeIssue, gitHubPRToForgePR } from "./mappers.js";
+import {
+  listIssuesImpl,
+  listPRsImpl,
+  getIssueImpl,
+  getPRImpl,
+  findPRByBranchImpl,
+  findPRsByBranchesImpl,
+  getCIStatusImpl,
+  findPRsByNumbersImpl,
+  findIssuesByNumbersImpl,
+  getCIStatusesImpl,
+  probeOpenPRListImpl,
+  getRepoMetadataImpl,
+  getReviewThreadsImpl,
+  getRateLimitImpl,
+  classifyPushErrorImpl,
+} from "./readOps.js";
+import {
+  createIssueImpl,
+  assignIssueImpl,
+  unassignIssueImpl,
+  createPRImpl,
+  closePRImpl,
+  reopenPRImpl,
+  mergePRImpl,
+  convertPRToDraftImpl,
+  markPRReadyForReviewImpl,
+  commentOnPRImpl,
+  editPRImpl,
+  closeIssueImpl,
+  reopenIssueImpl,
+  editIssueImpl,
+  addIssueCommentImpl,
+  addIssueLabelImpl,
+  removeIssueLabelImpl,
+  submitReviewImpl,
+  dismissReviewImpl,
+  requestReviewersImpl,
+} from "./mutations.js";
+
+export { gitHubIssueToForgeIssue, gitHubPRToForgePR };
 
 const reviewCapability: ReviewCapability = {
   getReviewThreads: getReviewThreadsImpl,
@@ -1772,28 +277,6 @@ const healthEventsCapability: HealthEventsCapability = {
   },
 };
 
-async function findIssuesByNumbersImpl(
-  repo: RepoRef,
-  issueNumbers: number[]
-): Promise<Map<number, Issue | null>> {
-  const items = await getIssuesByNumbersForContext(
-    { owner: repo.owner, repo: repo.repo },
-    issueNumbers
-  );
-  const result = new Map<number, Issue | null>();
-  for (const num of issueNumbers) {
-    result.set(num, null);
-  }
-  // The core's results align with its positive-integer-filtered input order;
-  // replay the same filter to pair numbers with items.
-  const valid = issueNumbers.filter((n) => typeof n === "number" && Number.isInteger(n) && n > 0);
-  valid.forEach((num, i) => {
-    const item = items[i];
-    result.set(num, item ? gitHubIssueToForgeIssue(item) : null);
-  });
-  return result;
-}
-
 export const githubForgeProvider: ForgeProviderImpl = {
   async getCredentials(): Promise<Credentials | null> {
     const token = GitHubAuth.getToken();
@@ -1895,396 +378,23 @@ export const githubForgeProvider: ForgeProviderImpl = {
     return `https://github.com/${repo.owner}/${repo.repo}/pull/${number}/files#diff-${hash}`;
   },
 
-  async createIssue(repo: RepoRef, input: CreateIssueInput): Promise<Issue> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    const title = input.title?.trim();
-    if (!title) {
-      throw new Error("Issue title is required.");
-    }
-    const requestBody: Record<string, unknown> = { title };
-    if (input.body) requestBody.body = input.body;
-    if (input.labels && input.labels.length > 0) requestBody.labels = input.labels;
-
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(requestBody),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      // 410 Gone = issues are disabled on the repository.
-      if (response.status === 410) {
-        throw new Error("Issues are disabled for this repository.");
-      }
-      throw new Error(
-        `Failed to create issue: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = (await response.json()) as Record<string, unknown>;
-    // Guard the mutation result: a malformed body must not surface as
-    // "Issue #undefined created" with an empty URL after caches are cleared.
-    if (typeof data.number !== "number" || typeof data.html_url !== "string") {
-      throw new Error("Unexpected response from GitHub: missing issue number or URL.");
-    }
-    // Provider-owned invalidation: drop list/stat caches so the new issue
-    // shows up in subsequent listIssues calls (callers no longer clear).
-    clearGitHubCaches();
-    return restToForgeIssue(data);
-  },
-
-  async assignIssue(repo: RepoRef, issueNumber: number, username: string): Promise<void> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}/assignees`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ assignees: [username] }),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to assign issue #${issueNumber} to ${username}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-  },
-
-  async unassignIssue(repo: RepoRef, issueNumber: number, username: string): Promise<void> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}/assignees`;
-    const response = await fetch(url, {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ assignees: [username] }),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to unassign issue #${issueNumber} from ${username}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-  },
-
-  async createPR(repo: RepoRef, input: CreatePRInput): Promise<PR> {
-    const token = requireGitHubToken();
-    const head = input.head?.trim();
-    const base = input.base?.trim();
-    const title = input.title?.trim();
-    if (!head) throw new Error("PR head branch is required.");
-    if (!base) throw new Error("PR base branch is required.");
-    if (!title) throw new Error("PR title is required.");
-    const requestBody: Record<string, unknown> = { head, base, title };
-    if (input.body) requestBody.body = input.body;
-    if (input.draft) requestBody.draft = true;
-
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: githubMutationHeaders(token),
-      body: JSON.stringify(requestBody),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      // 422 = head doesn't exist, no diff between head/base, or a PR for this
-      // head/base pair is already open. Surface GitHub's body so the user can tell.
-      throw new Error(
-        `Failed to create pull request: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = (await response.json()) as Record<string, unknown>;
-    if (typeof data.number !== "number" || typeof data.html_url !== "string") {
-      throw new Error("Unexpected response from GitHub: missing PR number or URL.");
-    }
-    clearPRCaches();
-    return restToForgePR(data);
-  },
-
-  async closePR(repo: RepoRef, prNumber: number): Promise<void> {
-    await patchPRState(repo, prNumber, "closed");
-    clearPRCaches();
-  },
-
-  async reopenPR(repo: RepoRef, prNumber: number): Promise<void> {
-    await patchPRState(repo, prNumber, "open");
-    clearPRCaches();
-  },
-
-  async mergePR(repo: RepoRef, prNumber: number, input?: MergePRInput): Promise<void> {
-    const token = requireGitHubToken();
-    const requestBody: Record<string, unknown> = {};
-    if (input?.mergeMethod) requestBody.merge_method = input.mergeMethod;
-    if (input?.commitTitle) requestBody.commit_title = input.commitTitle;
-    if (input?.commitMessage) requestBody.commit_message = input.commitMessage;
-
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}/merge`;
-    const response = await fetch(url, {
-      method: "PUT",
-      headers: githubMutationHeaders(token),
-      body: JSON.stringify(requestBody),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      // Classify by status, not message substring (lesson from #7079).
-      if (response.status === 405) {
-        throw new Error(
-          `Pull request #${prNumber} is not mergeable — it may be a draft, have conflicts, or fail required checks.`
-        );
-      }
-      if (response.status === 409) {
-        throw new Error(
-          `Pull request #${prNumber} head branch changed since it was last fetched — re-fetch the PR and try again.`
-        );
-      }
-      throw new Error(
-        `Failed to merge pull request #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    clearPRCaches();
-  },
-
-  async convertPRToDraft(repo: RepoRef, prNumber: number): Promise<void> {
-    // REST PATCH can't toggle draft state — GraphQL is the only path, and it
-    // needs the PR's node id, which the normalized PR type intentionally omits.
-    const nodeId = await fetchPRNodeId(repo, prNumber);
-    await dispatchQuery(CONVERT_PR_TO_DRAFT_MUTATION, { id: nodeId }, "convertPullRequestToDraft");
-    clearPRCaches();
-  },
-
-  async markPRReadyForReview(repo: RepoRef, prNumber: number): Promise<void> {
-    const nodeId = await fetchPRNodeId(repo, prNumber);
-    await dispatchQuery(
-      MARK_PR_READY_FOR_REVIEW_MUTATION,
-      { id: nodeId },
-      "markPullRequestReadyForReview"
-    );
-    clearPRCaches();
-  },
-
-  async commentOnPR(repo: RepoRef, prNumber: number, body: string): Promise<void> {
-    const token = requireGitHubToken();
-    // Reject an empty/whitespace body, but post the original text verbatim —
-    // trimming would mangle leading indentation in fenced code blocks.
-    if (!body?.trim()) throw new Error("Comment body is required.");
-    // PR comments use the issues endpoint — PR number === issue number on GitHub.
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${prNumber}/comments`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: githubMutationHeaders(token),
-      body: JSON.stringify({ body }),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to comment on pull request #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    // A comment changes the PR's comment count and timeline — drop PR caches so
-    // the next read reflects it (matches every other PR mutation).
-    clearPRCaches();
-  },
-
-  async editPR(repo: RepoRef, prNumber: number, input: EditPRInput): Promise<PR> {
-    const token = requireGitHubToken();
-    const requestBody: Record<string, unknown> = {};
-    if (typeof input.title === "string") {
-      const title = input.title.trim();
-      if (!title) throw new Error("PR title cannot be empty.");
-      requestBody.title = title;
-    }
-    if (typeof input.body === "string") requestBody.body = input.body;
-    if (Object.keys(requestBody).length === 0) {
-      throw new Error("Provide a title or body to edit.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/pulls/${prNumber}`;
-    const response = await fetch(url, {
-      method: "PATCH",
-      headers: githubMutationHeaders(token),
-      body: JSON.stringify(requestBody),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to edit pull request #${prNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = (await response.json()) as Record<string, unknown>;
-    if (typeof data.number !== "number" || typeof data.html_url !== "string") {
-      throw new Error("Unexpected response from GitHub: missing PR number or URL.");
-    }
-    clearPRCaches();
-    return restToForgePR(data);
-  },
-
-  async closeIssue(
-    repo: RepoRef,
-    issueNumber: number,
-    stateReason?: IssueCloseReason
-  ): Promise<Issue> {
-    const fields: Record<string, unknown> = { state: "closed" };
-    if (stateReason) fields.state_reason = stateReason;
-    return patchIssue(repo, issueNumber, fields, `Failed to close issue #${issueNumber}`);
-  },
-
-  async reopenIssue(repo: RepoRef, issueNumber: number): Promise<Issue> {
-    // Clear any prior `not_planned`/`completed` reason so a reopened issue isn't
-    // left with a stale close reason.
-    return patchIssue(
-      repo,
-      issueNumber,
-      { state: "open", state_reason: null },
-      `Failed to reopen issue #${issueNumber}`
-    );
-  },
-
-  async editIssue(repo: RepoRef, issueNumber: number, input: EditIssueInput): Promise<Issue> {
-    const fields: Record<string, unknown> = {};
-    if (input.title !== undefined) fields.title = input.title;
-    if (input.body !== undefined) fields.body = input.body;
-    if (Object.keys(fields).length === 0) {
-      throw new Error("editIssue requires at least one of title or body.");
-    }
-    return patchIssue(repo, issueNumber, fields, `Failed to edit issue #${issueNumber}`);
-  },
-
-  async addIssueComment(repo: RepoRef, issueNumber: number, body: string): Promise<IssueComment> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    if (!body || !body.trim()) {
-      throw new Error("Comment body is required.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}/comments`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ body }),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to comment on issue #${issueNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = (await response.json()) as Record<string, unknown>;
-    if (typeof data.id !== "number" || typeof data.html_url !== "string") {
-      throw new Error("Unexpected response from GitHub: missing comment id or URL.");
-    }
-    // A new comment changes the issue's comment count, which the tooltip cache
-    // holds. List/stat caches are unaffected, so a targeted invalidation is
-    // enough (avoids the heavier clearGitHubCaches()).
-    issueTooltipCache.invalidate(`${repo.owner}/${repo.repo}:${issueNumber}`);
-    return {
-      id: String(data.id),
-      body: typeof data.body === "string" ? data.body : "",
-      url: data.html_url,
-      author: restUserToForgeUser(data.user),
-      createdAt: isoToMs(data.created_at ?? data.updated_at),
-      rawData: data,
-    };
-  },
-
-  async addIssueLabel(repo: RepoRef, issueNumber: number, label: string): Promise<ForgeLabel[]> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    const name = label?.trim();
-    if (!name) {
-      throw new Error("Label name is required.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}/labels`;
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ labels: [name] }),
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to add label '${name}' to issue #${issueNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = await response.json();
-    clearGitHubCaches();
-    return restToForgeLabels(data);
-  },
-
-  async removeIssueLabel(repo: RepoRef, issueNumber: number, label: string): Promise<ForgeLabel[]> {
-    const token = GitHubAuth.getToken();
-    if (!token) {
-      throw new Error("GitHub token not configured. Set it in Settings.");
-    }
-    const name = label?.trim();
-    if (!name) {
-      throw new Error("Label name is required.");
-    }
-    const url = `https://api.github.com/repos/${repo.owner}/${repo.repo}/issues/${issueNumber}/labels/${encodeURIComponent(name)}`;
-    const response = await fetch(url, {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-      },
-      signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS),
-    });
-    if (response.status === 404) {
-      throw new Error(`Label '${name}' is not on issue #${issueNumber}.`);
-    }
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Failed to remove label '${name}' from issue #${issueNumber}: HTTP ${response.status}${text ? ` — ${text.slice(0, 200)}` : ""}`
-      );
-    }
-    const data = await response.json();
-    clearGitHubCaches();
-    return restToForgeLabels(data);
-  },
+  createIssue: createIssueImpl,
+  assignIssue: assignIssueImpl,
+  unassignIssue: unassignIssueImpl,
+  createPR: createPRImpl,
+  closePR: closePRImpl,
+  reopenPR: reopenPRImpl,
+  mergePR: mergePRImpl,
+  convertPRToDraft: convertPRToDraftImpl,
+  markPRReadyForReview: markPRReadyForReviewImpl,
+  commentOnPR: commentOnPRImpl,
+  editPR: editPRImpl,
+  closeIssue: closeIssueImpl,
+  reopenIssue: reopenIssueImpl,
+  editIssue: editIssueImpl,
+  addIssueComment: addIssueCommentImpl,
+  addIssueLabel: addIssueLabelImpl,
+  removeIssueLabel: removeIssueLabelImpl,
 
   async validateToken(token: string): Promise<AuthValidation> {
     if (!token || !token.trim()) {
